+++ test/parameters.js
@@ -1,0 +1,44 @@
+var tape = require( 'tape' );
+var util = require( 'util' );
+
+var importScript = require( '../import' );
+
+tape( 'interpretUserArgs() correctly handles arguments', function ( test ){
+  var testCases = [
+    [
+      [ '--deduplicate', '--admin-values', 'test'  ],
+      { deduplicate: true, adminValues: true, dirPath: 'test' },
+    ],
+    [
+      [ '--admin-values', 'test' ],
+      { deduplicate: false, adminValues: true, dirPath: 'test' },
+    ],
+    [
+      [ '--deduplicate', 'test' ],
+      { deduplicate: true, adminValues: false, dirPath: 'test' },
+    ]
+  ];
+
+  testCases.forEach( function execTestCase( testCase, ind ){
+    test.deepEqual(
+      importScript( testCase[ 0 ] ), testCase[ 1 ],
+      util.format( 'Arguments case %d passes.', ind )
+    );
+  });
+
+  var badArguments = [
+    [ 'not an arg', 'some dir' ],
+    [ '--deduplicate', 'not an arg', 'some dir' ],
+    [ '--deduplicate', 'not a dir' ],
+    [ '--deduplicate', 'package.json' ],
+  ];
+  badArguments.forEach( function execTestCase( testCase, ind ){
+    var errorObj = importScript( testCase );
+    test.ok(
+      'exitCode' in errorObj &&  'errMessage' in errorObj,
+      'Invalid arguments yield an error object: ' + ind
+    );
+  });
+  test.end();
+});
+
